#537 Separated tabs component
